First-time release the grid system
The grid system works like this:
we always treat base as the origin points,
and origin point should be the center points of the grid. For example for a 5x5 map, the orign point is at (2,2).

So we can use code to spawn cube or any objects on the map.